refactor: move batched_query into clients_utility
