+++ nisystemlink/clients/product/utilities/_constants.py
@@ -1,0 +1,2 @@
+class HttpConstants:
+    DEFAULT_QUERY_PRODUCTS_TAKE = 1000
